Tell git to track the branch we want to checkout
When the pull request targets a branch that isn't the default in the
repo the script was failing. This is because Travis does a shallow
clone, which means it only tracks the default branch (normally
`master`).

This change tells git to also track the branch the pull request targets,
which means when we perform the fetch it will also get a copy of that
branch.